Fix issue of password getting leaked
This commit changes the behaviour of the entrypoint Docker container, so
that it is not possible anymore to access the 'admin' account with `su
admin` and a password, since no password is then set anymore for the
admin account, i.e. the admin account is inaccessible from within the
account the user gets into the container.